Semantic keybindings
The 1, 2, and 3 keybindings for changing the page were conflicting with
Alt-1, Alt-2, and Alt-3 for changing tabs. I decided to make these more
semantic in addition to removing the conflict as well as adding two
additional keybindings. The new keybindings are as follows:

- t`A`gs
- ta`S`ks
- time `E`ntries
- `L`og in
- s`I`gn up

It was a conscious design decision that all of these are on or near the
home row, for additional comfort.