Update to Rails 5.1
Surely it can't be this easy. All I did is change the constraint on
the Rails version in `Gemfile` to `~> 5.1.0` and run `bundle update
rails`. I got one deprecation warning about halting callback chains on
returning false (a feature that I wasn't using anyways), so I just
removed the shim. I ran the tests and clicked through the application
and everything seems kosher.